requirements.txt
aiogram~=3.11.0
asyncpg~=0.29.0
pytest~=8.3.2
pytest-asyncio~=0.23.8
python-dotenv~=1.0.1
SQLAlchemy~=2.0.32